# api-spec/dist/openapi.yaml
openapi: 3.0.0
info:
  title: Hexlet Basics API
  description: |-
    Hexlet Basics API.

    Contract-first (ADR-0001): this TypeSpec compiles to `dist/openapi.yaml`
    (`pnpm spec`), which drives BOTH the Go server (ogen validators) and the TS
    frontend client + zod (hey-api, `pnpm generate`). Never hand-edit `dist/`.

    Naming: the domain concept is `Course`. The legacy Rails app called it
    `Language`; we keep `Language` ONLY in URL paths (e.g. `/languages`) for
    backward compatibility (ADR-0002), but every type/operation/schema uses
    `Course`. The full model/enum set lives in `models.tsp`.
  version: 0.0.0
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      description: Resolve the current user from the session cookie (for SSR).
      parameters:
        - name: Cookie
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        default:
          description: |-
            Default error response shared by every operation.

            `@error` emits an OpenAPI `default` response, keeping central transport
            failures typed without enumerating every status on every operation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Auth
...
  /session:
    post:
      operationId: createSession
      description: Log in with email + password; sets the JWT cookie.
      parameters: []
      responses:
        '200':
          description: |-
            A successful sign-in. Cookie headers are part of the HTTP contract so the
            implementation stays behind the generated ogen seam.
          headers:
            Set-Cookie:
              required: true
              schema:
                type: array
                items:
                  type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: |-
            Field-level validation errors, keyed by field name (each value is the list
            of messages for that field). Returned when a write fails validation —
            including constraints the schema cannot express, like uniqueness.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        default:
          description: |-
            Default error response shared by every operation.

            `@error` emits an OpenAPI `default` response, keeping central transport
            failures typed without enumerating every status on every operation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionInput'
    delete:
      operationId: deleteSession
      description: Log out; clears the JWT cookie.
      parameters: []
      responses:
        '204':
          description: A successful sign-out with the expired auth cookies.
          headers:
            Set-Cookie:
              required: true
              schema:
                type: array
                items:
                  type: string
        '401':
          description: The request is not authenticated (no/invalid session cookie).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        default:
          description: |-
            Default error response shared by every operation.

            `@error` emits an OpenAPI `default` response, keeping central transport
            failures typed without enumerating every status on every operation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Auth
      security:
        - UserSession: []
          XsrfToken: []
  /users:
    post:
      operationId: createUser
      description: Sign up; sets the JWT cookie on success.
      parameters: []
      responses:
        '201':
          description: A newly-created user and the auth cookies issued for that user.
          headers:
            Set-Cookie:
              required: true
              schema:
                type: array
                items:
                  type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: |-
            Field-level validation errors, keyed by field name (each value is the list
            of messages for that field). Returned when a write fails validation —
            including constraints the schema cannot express, like uniqueness.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        default:
          description: |-
            Default error response shared by every operation.

            `@error` emits an OpenAPI `default` response, keeping central transport
            failures typed without enumerating every status on every operation.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpInput'
components:
...
  schemas:
...
    CurrentUser:
      type: object
      required:
        - user
      properties:
        user:
          type: object
          allOf:
            - $ref: '#/components/schemas/User'
          nullable: true
      description: The current user resolved from the session cookie (null when anonymous).
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: URI identifying the problem type. `about:blank` means the HTTP status.
        title:
          type: string
          description: Short, localized summary for this problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code generated for this occurrence.
        detail:
          type: string
          description: Optional occurrence-specific explanation that is safe for clients.
        instance:
          type: string
          description: Optional URI identifying this particular occurrence.
      description: |-
        RFC 9457 problem details returned by the transport when request decoding,
        authorization, persistence, or an unexpected server failure prevents an
        operation from producing its declared domain response.
...
    SessionInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      description: Email + password login.
    SignUpInput:
      type: object
      required:
        - firstName
        - email
        - password
      properties:
        firstName:
          type: string
          nullable: true
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
      description: 'Sign-up form (legacy: `UserSignUpForm`).'
...
    User:
      type: object
      required:
        - id
        - firstName
        - lastName
        - name
        - email
        - admin
        - canAccessAdmin
        - assistantMessagesCount
        - createdAt
        - createdAtAsTimestamp
        - type
      properties:
        id:
          type: integer
          format: int32
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        admin:
          type: boolean
          nullable: true
        canAccessAdmin:
          type: boolean
        assistantMessagesCount:
          type: integer
          format: int32
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdAtAsTimestamp:
          type: integer
          format: int64
          nullable: true
        type:
          type: string
          enum:
            - user
      description: 'A user (legacy: `User`). Passwords are never part of a response.'
...
    ValidationError:
      type: object
      required:
        - errors
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      description: |-
        Field-level validation errors, keyed by field name (each value is the list
        of messages for that field). Returned when a write fails validation —
        including constraints the schema cannot express, like uniqueness.
  securitySchemes:
    UserSession:
      type: apiKey
      in: cookie
      name: JWT
    XsrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
    AdminSession:
      type: apiKey
      in: cookie
      name: JWT